# Greatly speed up reading/writing slices with #[inline]