@cmr 2015-03-25
@bors: r+ rollup