@Gankro 2015-03-25
Seriously if basically all code should be #[inline] should we be considering cranking up how aggressively we hint to LLVM to inline things?