@erickt 2015-03-25
@bors: r=cmr rollup