@erickt 2015-04-01
@Gankro: Doesn't llvm need the functions in the metadata to do inlining across crates? Assuming of course we're not using LTO.